Move checksum and get_variable_values_sorted from widget.utils.scaling to widget.utils.
In the same commit, I removed get_variable_value_indices that is still used in several functions in scaling, but will be removed in a future commit. Please inform me if I should invest some more time in splitting this commit in two.
