+++ Orange/widgets/utils/__init__.py
@@ -37,3 +37,25 @@ def to_html(str):
         replace("<", "&#60;").replace(">", "&#62;").replace("=\\=", "&#8800;")
 
 getHtmlCompatibleString = to_html
+
+
+def checksum(x):
+    if x is None:
+        return None
+    try:
+        return x.checksum()
+    except:
+        return float('nan')
+
+
+def get_variable_values_sorted(variable):
+    """
+    Return a list of sorted values for given attribute, if all its values can be
+    cast to int's.
+    """
+    if variable.is_continuous:
+        return []
+    try:
+        return sorted(variable.values, key=int)
+    except ValueError:
+        return variable.values
